Rename file to view
